# Conflicts:
#	.idea/modules.xml
#	app/src/main/AndroidManifest.xml
#	app/src/main/java/ca/uottawa/cookingwithgarzon/CreateRecipe.java
#	app/src/main/res/layout/activity_cart_ingredient.xml
#	app/src/main/res/values/strings.xml

+++ app/src/main/java/ca/uottawa/cookingwithgarzon/model/MealType.java
@@ -8,7 +8,7 @@ public class MealType {
     private long _id;
     private String _name;
 
-    public int get_id() {
+    public long get_id() {
         return _id;
     }
